pyproject.toml
[project]
name = "10701-light-estimation"
version = "0.1.0"
description = "Near-field light source estimation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mathutils>=3.3.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pillow>=9.0.0",
    "mitsuba>=3.7.1",
    "polyscope>=2.5.0",
    "trimesh>=4.0.0",
    "matplotlib>=3.10.7",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]

[tool.ruff]
target-version = "py311"
line-length = 89
fix = true
extend-select = ["I"] # sort imports
